flask_service/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.22
psycopg2-binary==2.9.7
python-dotenv==1.0.1
requests==2.31.0
Werkzeug==2.3.7
python-multipart==0.0.6
flask-cors==3.0.10